fix(vaccination-reminder): can not inject websocket to api request
