+++ packages/facility-server/app/routes/apiv1/telegram/telegramRoutes.js
@@ -13,8 +13,8 @@ export const telegramRoutes = express.Router();
 const getTelegramBotInfo = async ws => {
   return await new Promise((resolve, reject) => {
     try {
-      ws.emit(WS_EVENTS.TELEGRAM_GET_BOT_INFO);
-      ws.listenOnce(WS_EVENTS.TELEGRAM_BOT_INFO, botInfo => resolve(botInfo));
+      ws?.emit(WS_EVENTS.TELEGRAM_GET_BOT_INFO);
+      ws?.listenOnce(WS_EVENTS.TELEGRAM_BOT_INFO, botInfo => resolve(botInfo));
     } catch (e) {
       reject(e);
     }
@@ -25,7 +25,7 @@ telegramRoutes.get(
   '/bot-info',
   expressAsyncHandler(async (req, res) => {
     req.flagPermissionChecked();
-    const botInfo = await getTelegramBotInfo(req.websocketClientService);
+    const botInfo = await getTelegramBotInfo(req?.websocketClientService);
     res.send(botInfo);
   }),
 );
