+++ packages/facility-server/app/services/websocketClientService.js
@@ -23,7 +23,7 @@ export const defineWebsocketClientService = injector => {
       contact.connectionDetails = { chatId };
       await contact.save();
 
-      injector.websocketService.emit(WS_EVENTS.TELEGRAM_SUBSCRIBE_SUCCESS, { contactId, chatId });
+      injector.websocketService?.emit(WS_EVENTS.TELEGRAM_SUBSCRIBE_SUCCESS, { contactId, chatId });
     },
   );
 
@@ -38,7 +38,7 @@ export const defineWebsocketClientService = injector => {
       if (!contact) return;
 
       await contact.destroy();
-      injector.websocketService.emit(WS_EVENTS.TELEGRAM_UNSUBSCRIBE_SUCCESS, { contactId });
+      injector.websocketService?.emit(WS_EVENTS.TELEGRAM_UNSUBSCRIBE_SUCCESS, { contactId });
     },
   );
